+++ CHANGES.rst
@@ -37,3 +37,6 @@ Bug Fixes:
 
 - Add zcml-condition plone-52 for conditional configuration.
   [pbauer]
+
+- Use getSite in set_own_login_name to get the portals acl_users.
+  [pbauer]